gymos-backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ─── Database ────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ─── Validation / Serialization ──────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── Authentication / Security ───────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ─── HTTP / Async Utilities ──────────────────────────────────────
httpx==0.27.0
anyio==4.3.0

# ─── Environment ─────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Testing ─────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0

# ─── Monitoring / Logging ────────────────────────────────────────
structlog==24.1.0

# ─── Dev Tools ───────────────────────────────────────────────────
black==24.4.2
ruff==0.4.4
mypy==1.10.0
